trappy/ftrace: add support to account line numbers
Knowing the line number of a trace event into the original trace file can
be super useful for two main reasons:
1. it can be used as a truly unique ID of each DF entry, since it is not
   subject to the possible index collisions due to the timestamp resoluton
2. it can be super-useful to correlate DF entries with the original trace

If this ID is also tracked the same way tools like kernelshark does, than
we have a definitive and easy tool to match DF records with kernelshark.

Since kernelshark uses a 0-based indexing for trace events, let's account
for proper trace events line numbers starting from 0. This number is available
as a class property, in case we are interested to know to real total number
of events which are available in the original trace. Moreover, it will be
used by a following patch to add a unique ID to each DF entry.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>